navigation zwischen den einzelnen Seiten vollständig Logo verkleinert. html tags zur optimalen Textdarstellung hinzugefuegt.

+++ Shift/src/webservices/GmWS.java
@@ -117,6 +117,6 @@ public interface GmWS {
 	 */
 
 	@WebMethod
-	int sendOrder(String name, String apartmentID);
+	long sendOrder(String name, String apartmentID);
 
 }
